+++ Assets/Scripts/Global.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f94cc0cc2e232384c8c99727079bf4f3
+folderAsset: yes
+timeCreated: 1460899076
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
