+++ Assets/Scripts/Global/EnumTypes.cs
