+++ Assets/Scripts/Interaction.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 52bd9d2ff6c87364793833a648eb6a64
+folderAsset: yes
+timeCreated: 1460899076
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
