+++ Assets/Scripts/Interaction/Blessable.cs
