+++ Assets/Scripts/Interaction/Blessable.cs.meta
