+++ Assets/Scripts/Interaction/Diggable.cs.meta
