+++ Assets/Scripts/Interaction/NoopRangedInteraction.cs
