+++ Assets/Scripts/Interaction/Readable.cs.meta
