+++ Assets/Scripts/Interaction/SimpleDestructionRangedInteraction.cs
