+++ Assets/Scripts/Platforming.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 5a3934880784a9c489a99316996ef9a9
+folderAsset: yes
+timeCreated: 1460899076
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
