+++ Assets/Scripts/Player.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 8fcf196bb7dec6e43949a13041855f8a
+folderAsset: yes
+timeCreated: 1460899076
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
